+++ lib/oauth/request_proxy.rb
@@ -2,24 +2,28 @@
 
 module OAuth
   module RequestProxy
-    def self.available_proxies # :nodoc:
-      @available_proxies ||= {}
-    end
+    AVAILABLE_PROXIES = {}
 
-    def self.proxy(request, options = {})
-      return request if request.is_a?(OAuth::RequestProxy::Base)
-
-      klass = available_proxies[request.class]
-
-      # Search for possible superclass matches.
-      if klass.nil?
-        request_parent = available_proxies.keys.find { |rc| request.is_a?(rc) }
-        klass = available_proxies[request_parent]
+    class << self
+      def available_proxies # :nodoc:
+        AVAILABLE_PROXIES
       end
 
-      raise UnknownRequestType, request.class.to_s unless klass
+      def proxy(request, options = {})
+        return request if request.is_a?(OAuth::RequestProxy::Base)
 
-      klass.new(request, options)
+        klass = available_proxies[request.class]
+
+        # Search for possible superclass matches.
+        if klass.nil?
+          request_parent = available_proxies.keys.find { |rc| request.is_a?(rc) }
+          klass = available_proxies[request_parent]
+        end
+
+        raise UnknownRequestType, request.class.to_s unless klass
+
+        klass.new(request, options)
+      end
     end
 
     class UnknownRequestType < RuntimeError; end